requirements.txt
python-dateutil
gtfs-realtime-bindings
urwid

requests
click
haversine